+++ packages/desktop/package.json
@@ -29,6 +29,10 @@
     },
     "dependencies": {
         "@amplitude/analytics-node": "^1.1.3",
+        "@ethereumjs/common": "^3.1.2",
+        "@ethereumjs/rlp": "^4.0.1",
+        "@ethereumjs/tx": "^4.1.2",
+        "@ethereumjs/util": "^8.0.6",
         "@ledgerhq/hw-app-eth": "^6.32.2",
         "@ledgerhq/hw-transport-node-hid": "^6.27.13",
         "@ledgerhq/logs": "^6.10.1",
