+++ packages/shared/lib/core/ledger/actions/index.ts
@@ -3,5 +3,4 @@ export * from './displayNotificationForLedgerProfile'
 export * from './getLedgerDeviceStatus'
 export * from './pollLedgerNanoStatus'
 export * from './promptUserToConnectLedger'
-export * from './registerLayer2EventHandlers'
 export * from './registerLedgerDeviceEventHandlers'
